faster

+++ virtual/src/main/java/com/fpetrola/z80/transformations/InstructionTransformer.java
@@ -236,7 +236,7 @@ public class InstructionTransformer<T extends WordNumber> extends InstructionTra
   }
 
   public void visitingRst(RST<T> rst) {
-    setCloned(instructionFactory.RST(rst.getP()), rst);
+    setCloned(instructionFactory.RST(rst.getP().intValue()), rst);
   }
 
   public void visitPush(Push push) {
